* Improved performance and reduced loading time (2 years old code formatting) * Added some custom donator statue/companions + improved code
Bugfixes:
* Fixed Abyssal Blade and Skull Basher passive bashes working on Faceless Void, Slardar, Spirit Breaker and Troll Warlord
* Fixed Jarnbjorn proc full cd when cut a tree down (now 2.5s)
* Fixed Mirana's ultimate not hiding companion
* Fixed Smoke of Deceit not hiding companion
