+++ game/dota_addons/dota_imba/scripts/vscripts/battlepass/donator.lua
@@ -3,9 +3,7 @@ function DonatorStatue(ID, statue_info)
 		statue_info = IMBA_DONATOR_STATUE[tostring(PlayerResource:GetSteamID(ID))]
 	end
 
-	print(ID)
-	PrintTable(statue_info)
-
+	local pedestal_name = "npc_imba_donator_pedestal"
 	local hero = PlayerResource:GetSelectedHeroEntity(ID)
 --	if hero.donator_statue then
 --		hero.donator_statue:ForceKill(false)
@@ -20,6 +18,11 @@ function DonatorStatue(ID, statue_info)
 --	end
 
 	local team = "good"
+
+	if PlayerResource:GetPlayer(ID):GetTeam() == 3 then
+		team = "bad"
+	end
+
 	local fillers = {
 		team.."_filler_2",
 		team.."_filler_4",
@@ -27,20 +30,12 @@ function DonatorStatue(ID, statue_info)
 		team.."_filler_7",
 	}
 
-	if PlayerResource:GetPlayer(ID):GetTeam() == 3 then
-		team = "bad"
-	end
-
 	for _, ent_name in pairs(fillers) do
 		local filler = Entities:FindByName(nil, ent_name)
 		if filler then
 			local abs = filler:GetAbsOrigin()
 
 			filler:RemoveSelf()
-
---			print(statue_info)
---			print(statue_info[1])
---			print(statue_info[2])
 
 			local unit = CreateUnitByName(statue_info[2], abs, true, nil, nil, PlayerResource:GetPlayer(ID):GetTeam())
 			unit:SetModelScale(statue_info[1])
@@ -54,31 +49,47 @@ function DonatorStatue(ID, statue_info)
 
 			if api.imba.is_donator(steam_id) == 1 then
 				unit:SetCustomHealthLabel(name, 160, 20, 20)
---			elseif api.imba.is_donator(steam_id) == 2 then
---				unit:SetCustomHealthLabel("Baumi", 160, 20, 20)
+				pedestal_name = "npc_imba_donator_pedestal_cookies"
+			elseif api.imba.is_donator(steam_id) == 2 then
+				unit:SetCustomHealthLabel("sutherncuck", 0, 204, 255)
+				pedestal_name = "npc_imba_donator_pedestal_developer_"..team
 			elseif api.imba.is_donator(steam_id) == 3 then
 				unit:SetCustomHealthLabel(name, 160, 20, 20)
 			elseif api.imba.is_donator(steam_id) == 4 then
 				unit:SetCustomHealthLabel(name, 240, 50, 50)
+				pedestal_name = "npc_imba_donator_pedestal_ember_"..team
 			elseif api.imba.is_donator(steam_id) == 5 then
 				unit:SetCustomHealthLabel(name, 218, 165, 32)
+				pedestal_name = "npc_imba_donator_pedestal_golden_"..team
 			elseif api.imba.is_donator(steam_id) == 7 then
 				unit:SetCustomHealthLabel(name, 47, 91, 151)
+				pedestal_name = "npc_imba_donator_pedestal_salamander_"..team
 			elseif api.imba.is_donator(steam_id) == 8 then
 				unit:SetCustomHealthLabel(name, 153, 51, 153)
-			elseif api.imba.is_donator(steam_id) then
+				pedestal_name = "npc_imba_donator_pedestal_icefrog"
+			elseif api.imba.is_donator(steam_id) then -- 6: donator, 0: lesser donator
 				unit:SetCustomHealthLabel(name, 45, 200, 45)
 			end
 
 			if statue_info[2] == "npc_imba_donator_statue_suthernfriend" then
 				unit:SetMaterialGroup("1")
+			elseif statue_info[2] == "npc_imba_donator_statue_tabisama" then
+				unit:SetAbsOrigin(unit:GetAbsOrigin() + Vector(0, 0, 40))
 			end
 
-			local pedestal = CreateUnitByName("npc_imba_donator_statue", abs, true, nil, nil, PlayerResource:GetPlayer(ID):GetTeam())
+			if statue_info[2] == "npc_imba_donator_statue_zonnoz" then
+				pedestal_name = "npc_imba_donator_pedestal_pudge_arcana"
+			end
+
+			local pedestal = CreateUnitByName(pedestal_name, abs, true, nil, nil, PlayerResource:GetPlayer(ID):GetTeam())
 			pedestal:AddNewModifier(pedestal, nil, "modifier_imba_contributor_statue", {})
 --			pedestal:SetModelScale(statue_info[1])
 			pedestal:SetAbsOrigin(abs + Vector(0, 0, 45))
 			unit.pedestal = pedestal
+
+			if statue_info[2] == "npc_imba_donator_statue_zonnoz" then
+				pedestal:SetMaterialGroup("1")
+			end
 
 			return
 		end
@@ -86,11 +97,17 @@ function DonatorStatue(ID, statue_info)
 end
 
 function DonatorCompanion(ID, unit_name)
+if IMBA_DONATOR_COMPANION[tostring(PlayerResource:GetSteamID(ID))] then 
+	unit_name = IMBA_DONATOR_COMPANION[tostring(PlayerResource:GetSteamID(ID))]
+end
+
 if unit_name == nil then return end
 local hero = PlayerResource:GetPlayer(ID):GetAssignedHero()
 local color = hero:GetFittingColor()
 local model = GetKeyValueByHeroName(unit_name, "Model")
 local model_scale = GetKeyValueByHeroName(unit_name, "ModelScale")
+
+--	print(unit_name, model, model_scale)
 
 	if hero.companion then
 		hero.companion:ForceKill(false)
