+++ game/dota_addons/dota_imba/scripts/vscripts/items/item_abyssal_blade.lua
@@ -28,41 +28,41 @@ LinkLuaModifier("modifier_imba_abyssal_blade_skull_crash", "items/item_abyssal_b
 LinkLuaModifier("modifier_imba_abyssal_blade_skull_break", "items/item_abyssal_blade", LUA_MODIFIER_MOTION_NONE)
 
 function item_imba_abyssal_blade:GetIntrinsicModifierName()
-    return "modifier_imba_abyssal_blade"
+	return "modifier_imba_abyssal_blade"
 end
 
 function item_imba_abyssal_blade:OnSpellStart()
-    -- Ability properties
-    local caster = self:GetCaster()
-    local ability = self
-    local target = self:GetCursorTarget()
-    local sound_cast = "DOTA_Item.AbyssalBlade.Activate"    
-    local particle_abyssal = "particles/items_fx/abyssal_blade.vpcf"
-    local modifier_bash = "modifier_imba_abyssal_blade_bash"
-    local modifier_break = "modifier_imba_abyssal_blade_skull_break"
+	-- Ability properties
+	local caster = self:GetCaster()
+	local ability = self
+	local target = self:GetCursorTarget()
+	local sound_cast = "DOTA_Item.AbyssalBlade.Activate"    
+	local particle_abyssal = "particles/items_fx/abyssal_blade.vpcf"
+	local modifier_bash = "modifier_imba_abyssal_blade_bash"
+	local modifier_break = "modifier_imba_abyssal_blade_skull_break"
 
-    -- Ability specials
-    local active_stun_duration = ability:GetSpecialValueFor("active_stun_duration")
+	-- Ability specials
+	local active_stun_duration = ability:GetSpecialValueFor("active_stun_duration")
 	local actual_break_duration = ability:GetSpecialValueFor("actual_break_duration")
 
-    -- Play cast sound
-    EmitSoundOn(sound_cast, target)
-    
-    -- If the target possesses a ready Linken's Sphere, do nothing
-    if target:GetTeamNumber() ~= caster:GetTeamNumber() then
-        if target:TriggerSpellAbsorb(ability) then
-            return nil
-        end
-    end
+	-- Play cast sound
+	EmitSoundOn(sound_cast, target)
+	
+	-- If the target possesses a ready Linken's Sphere, do nothing
+	if target:GetTeamNumber() ~= caster:GetTeamNumber() then
+		if target:TriggerSpellAbsorb(ability) then
+			return nil
+		end
+	end
 
-    -- Add particle effect
-    local particle_abyssal_fx = ParticleManager:CreateParticle(particle_abyssal, PATTACH_ABSORIGIN_FOLLOW, target)
-    ParticleManager:SetParticleControl(particle_abyssal_fx, 0, target:GetAbsOrigin())
-    ParticleManager:ReleaseParticleIndex(particle_abyssal_fx)
+	-- Add particle effect
+	local particle_abyssal_fx = ParticleManager:CreateParticle(particle_abyssal, PATTACH_ABSORIGIN_FOLLOW, target)
+	ParticleManager:SetParticleControl(particle_abyssal_fx, 0, target:GetAbsOrigin())
+	ParticleManager:ReleaseParticleIndex(particle_abyssal_fx)
 
-    -- Stun and break the target for the duration
-    target:AddNewModifier(caster, ability, modifier_bash, {duration = active_stun_duration})
-    target:AddNewModifier(caster, ability, modifier_break, {duration = actual_break_duration})
+	-- Stun and break the target for the duration
+	target:AddNewModifier(caster, ability, modifier_bash, {duration = active_stun_duration})
+	target:AddNewModifier(caster, ability, modifier_break, {duration = actual_break_duration})
 end
 
 
@@ -76,64 +76,64 @@ function modifier_imba_abyssal_blade:IsPermanent() return true end
 function modifier_imba_abyssal_blade:GetAttributes() return MODIFIER_ATTRIBUTE_MULTIPLE end 
 
 function modifier_imba_abyssal_blade:OnCreated()
-    -- Ability properties
-    self.caster = self:GetCaster()
-    self.ability = self:GetAbility()
-    self.parent = self:GetParent()
-    self.modifier_self = "modifier_imba_abyssal_blade"
-    self.modifier_unique = "modifier_imba_abyssal_blade_unique"
+	-- Ability properties
+	self.caster = self:GetCaster()
+	self.ability = self:GetAbility()
+	self.parent = self:GetParent()
+	self.modifier_self = "modifier_imba_abyssal_blade"
+	self.modifier_unique = "modifier_imba_abyssal_blade_unique"
 
-    -- Ability specials
-    self.bonus_damage = self.ability:GetSpecialValueFor("bonus_damage")
-    self.bonus_strength = self.ability:GetSpecialValueFor("bonus_strength")   
-    self.bonus_hp_regen = self.ability:GetSpecialValueFor("bonus_hp_regen") 
-    self.bonus_health = self.ability:GetSpecialValueFor("bonus_health")
-    self.damage_block = self.ability:GetSpecialValueFor("damage_block")
+	-- Ability specials
+	self.bonus_damage = self.ability:GetSpecialValueFor("bonus_damage")
+	self.bonus_strength = self.ability:GetSpecialValueFor("bonus_strength")   
+	self.bonus_hp_regen = self.ability:GetSpecialValueFor("bonus_hp_regen") 
+	self.bonus_health = self.ability:GetSpecialValueFor("bonus_health")
+	self.damage_block = self.ability:GetSpecialValueFor("damage_block")
 
-    if IsServer() then
-        -- Grant the unique modifier for bashing
-        if not self.caster:HasModifier(self.modifier_unique) then
-            self.caster:AddNewModifier(self.caster, self.ability, self.modifier_unique, {})
-        end
-    end
+	if IsServer() then
+		-- Grant the unique modifier for bashing
+		if not self.caster:HasModifier(self.modifier_unique) then
+			self.caster:AddNewModifier(self.caster, self.ability, self.modifier_unique, {})
+		end
+	end
 end
 
 function modifier_imba_abyssal_blade:OnDestroy()
-    if IsServer() then
-        -- If there are no more skull bashers in inventory, remove the skull basher unique modifier altogether
-        if not self.caster:HasModifier(self.modifier_self) then
-            self.caster:RemoveModifierByName(self.modifier_unique)
-        end
-    end
+	if IsServer() then
+		-- If there are no more skull bashers in inventory, remove the skull basher unique modifier altogether
+		if not self.caster:HasModifier(self.modifier_self) then
+			self.caster:RemoveModifierByName(self.modifier_unique)
+		end
+	end
 end
 
 function modifier_imba_abyssal_blade:DeclareFunctions()
-    local decFuncs = {MODIFIER_PROPERTY_STATS_STRENGTH_BONUS,
-                      MODIFIER_PROPERTY_PREATTACK_BONUS_DAMAGE,
-                      MODIFIER_PROPERTY_HEALTH_BONUS,
-                      MODIFIER_PROPERTY_HEALTH_REGEN_CONSTANT}
+	local decFuncs = {MODIFIER_PROPERTY_STATS_STRENGTH_BONUS,
+					  MODIFIER_PROPERTY_PREATTACK_BONUS_DAMAGE,
+					  MODIFIER_PROPERTY_HEALTH_BONUS,
+					  MODIFIER_PROPERTY_HEALTH_REGEN_CONSTANT}
 
-    return decFuncs
+	return decFuncs
 end
 
 function modifier_imba_abyssal_blade:GetModifierBonusStats_Strength()
-    return self.bonus_strength
+	return self.bonus_strength
 end
 
 function modifier_imba_abyssal_blade:GetModifierPreAttack_BonusDamage()
-    return self.bonus_damage
+	return self.bonus_damage
 end
 
 function modifier_imba_abyssal_blade:GetModifierHealthBonus()
-    return self.bonus_health
+	return self.bonus_health
 end
 
 function modifier_imba_abyssal_blade:GetModifierConstantHealthRegen()
-    return self.bonus_hp_regen
+	return self.bonus_hp_regen
 end
 
 function modifier_imba_abyssal_blade:GetCustomDamageBlockUnique()
-    return self.damage_block
+	return self.damage_block
 end
 
 
@@ -147,138 +147,134 @@ function modifier_imba_abyssal_blade_unique:IsDebuff() return false end
 function modifier_imba_abyssal_blade_unique:IsPermanent() return true end
 
 function modifier_imba_abyssal_blade_unique:OnCreated()
-    -- Ability properties
-    self.caster = self:GetCaster()
-    self.ability = self:GetAbility()
-    self.parent = self:GetParent()
-    self.modifier_bash = "modifier_imba_abyssal_blade_bash"
-    self.modifier_skull_crash = "modifier_imba_abyssal_blade_skull_crash"
-    self.modifier_skull_break = "modifier_imba_abyssal_blade_skull_break"
-    self.modifier_internal_cd = "modifier_imba_abyssal_blade_internal_cd"
+	-- Ability properties
+	self.caster = self:GetCaster()
+	self.ability = self:GetAbility()
+	self.parent = self:GetParent()
+	self.modifier_bash = "modifier_imba_abyssal_blade_bash"
+	self.modifier_skull_crash = "modifier_imba_abyssal_blade_skull_crash"
+	self.modifier_skull_break = "modifier_imba_abyssal_blade_skull_break"
+	self.modifier_internal_cd = "modifier_imba_abyssal_blade_internal_cd"
 
-    -- Ability specials
-    self.bonus_range_melee = self.ability:GetSpecialValueFor("bonus_range_melee")
-    self.bash_chance_melee = self.ability:GetSpecialValueFor("bash_chance_melee")
-    self.bash_chance_ranged = self.ability:GetSpecialValueFor("bash_chance_ranged")
-    self.stun_duration = self.ability:GetSpecialValueFor("stun_duration")
-    self.bash_damage = self.ability:GetSpecialValueFor("bash_damage")
-    self.skull_break_duration = self.ability:GetSpecialValueFor("skull_break_duration")
-    self.internal_bash_cd = self.ability:GetSpecialValueFor("internal_bash_cd")
-    self.insta_skull_break_chance_pct = self.ability:GetSpecialValueFor("insta_skull_break_chance_pct")
+	-- Ability specials
+	self.bonus_range_melee = self.ability:GetSpecialValueFor("bonus_range_melee")
+	self.bash_chance_melee = self.ability:GetSpecialValueFor("bash_chance_melee")
+	self.bash_chance_ranged = self.ability:GetSpecialValueFor("bash_chance_ranged")
+	self.stun_duration = self.ability:GetSpecialValueFor("stun_duration")
+	self.bash_damage = self.ability:GetSpecialValueFor("bash_damage")
+	self.skull_break_duration = self.ability:GetSpecialValueFor("skull_break_duration")
+	self.internal_bash_cd = self.ability:GetSpecialValueFor("internal_bash_cd")
+	self.insta_skull_break_chance_pct = self.ability:GetSpecialValueFor("insta_skull_break_chance_pct")
 	self.actual_break_duration = self.ability:GetSpecialValueFor("actual_break_duration")
 end
 
 function modifier_imba_abyssal_blade_unique:DeclareFunctions()
-    local decFuncs = {MODIFIER_EVENT_ON_ATTACK_LANDED,
-                      MODIFIER_PROPERTY_ATTACK_RANGE_BONUS,
-                      MODIFIER_EVENT_ON_ATTACK,
-                      MODIFIER_PROPERTY_PROCATTACK_BONUS_DAMAGE_MAGICAL}
+	local decFuncs = {MODIFIER_EVENT_ON_ATTACK_LANDED,
+					  MODIFIER_PROPERTY_ATTACK_RANGE_BONUS,
+					  MODIFIER_EVENT_ON_ATTACK,
+					  MODIFIER_PROPERTY_PROCATTACK_BONUS_DAMAGE_MAGICAL}
 
-    return decFuncs
+	return decFuncs
 end
 
 function modifier_imba_abyssal_blade_unique:GetModifierAttackRangeBonus()
-    if not self.caster:IsRangedAttacker() then
-        return self.bonus_range_melee
-    end
+	if not self.caster:IsRangedAttacker() then
+		return self.bonus_range_melee
+	end
 end
 
 function modifier_imba_abyssal_blade_unique:OnAttack(keys)
-    if IsServer() then
-        local attacker = keys.attacker
-        local target = keys.target
+	if IsServer() then
+		local attacker = keys.attacker
+		local target = keys.target
 
-        -- Only apply if the attacker is the caster
-        if attacker == self.caster then
+		-- Only apply if the attacker is the caster
+		if attacker == self.caster then
 
-            -- Check if this is a ranged or melee attacker
-            local chance
-            if self.caster:IsRangedAttacker() then
-                chance = self.bash_chance_ranged
-            else
-                chance = self.bash_chance_melee
-            end
+			-- Check if this is a ranged or melee attacker
+			local chance
+			if self.caster:IsRangedAttacker() then
+				chance = self.bash_chance_ranged
+			else
+				chance = self.bash_chance_melee
+			end
 
-            -- If the attacker is an illusion, do nothing
-            if self.caster:IsIllusion() then
-                return nil
-            end
+			-- If the attacker is an illusion, do nothing
+			if self.caster:IsIllusion() then
+				return nil
+			end
 
-            -- If the target is a building, do nothing
-            if target:IsBuilding() then
-                return nil                
-            end
+			-- If the target is a building, do nothing
+			if target:IsBuilding() then
+				return nil                
+			end
 
-            if attacker:GetTeamNumber() == target:GetTeamNumber() then
-                return nil
-            end
+			if attacker:GetTeamNumber() == target:GetTeamNumber() then
+				return nil
+			end
 
-            -- If the Abyssal Blade is on cooldown (internal), do nothing
-            if self.caster:HasModifier(self.modifier_internal_cd) then                
-                return nil
-            end
+			-- If the Abyssal Blade is on cooldown (internal), do nothing
+			if self.caster:HasModifier(self.modifier_internal_cd) then                
+				return nil
+			end
 
-            -- Roll for a bash!
-            if RollPseudoRandom(chance, self) then                
-                self.bash_proc = true
-            end
-        end
-    end
+			-- Roll for a bash!
+			if RollPseudoRandom(chance, self) then                
+				self.bash_proc = true
+			end
+		end
+	end
 end
 
 function modifier_imba_abyssal_blade_unique:OnAttackLanded(keys)
-    if IsServer() then
-        local attacker = keys.attacker
-        local target = keys.target
+	if IsServer() then
+		local attacker = keys.attacker
+		local target = keys.target
 
-        -- If the attacker is one of the forbidden heroes, do nothing
-        local forbidden_heroes = {npc_dota_hero_slardar = true,
-                                  npc_dota_hero_faceless_void = true,
-                                  npc_dota_hero_spirit_breaker = true,
-                                  npc_dota_hero_troll_warlord = true}
+		-- If the attacker is one of the forbidden heroes, do nothing
+		for _, restricted_hero in pairs(IMBA_DISABLED_SKULL_BASHER) do
+			if restricted_hero == attacker:GetUnitName() then
+				return nil
+			end
+		end
 
-        if forbidden_heroes[self.caster:GetUnitName()] then
-            return nil
-        end
+		-- Only apply if the caster is the attacker
+		if attacker == self.caster then
+			if self.bash_proc then
+				-- Bash bash bash
+				target:AddNewModifier(self.caster, self.ability, self.modifier_bash, {duration = self.stun_duration})
 
-        -- Only apply if the caster is the attacker
-        if attacker == self.caster then            
+				-- Make the ability go into an internal cooldown
+				self.caster:AddNewModifier(self.caster, self.ability, self.modifier_internal_cd, {duration = self.internal_bash_cd})
 
-            if self.bash_proc then
-                -- Bash bash bash
-                target:AddNewModifier(self.caster, self.ability, self.modifier_bash, {duration = self.stun_duration})
+				-- If the target is not skull crashed yet, try to immediately CRUSH IT!
+				if not target:HasModifier(self.modifier_skull_crash) then
+					if RollPseudoRandom(self.insta_skull_break_chance_pct, self) then
+						target:AddNewModifier(self.caster, self.ability, self.modifier_skull_break, {duration = self.actual_break_duration})  
+					else
+						target:AddNewModifier(self.caster, self.ability, self.modifier_skull_crash, {duration = self.skull_break_duration})
+					end
+				else
+					-- Otherwise, it was ALREADY CRUSHED! BREAK IT!!!!!!!!!!!! BREAK IT!!!!!!!!!!!!!!!
+					-- Consume skull crash modifier
+					target:RemoveModifierByName(self.modifier_skull_crash)
 
-                -- Make the ability go into an internal cooldown
-                self.caster:AddNewModifier(self.caster, self.ability, self.modifier_internal_cd, {duration = self.internal_bash_cd})
+					-- Apply BREAK!
+					target:AddNewModifier(self.caster, self.ability, self.modifier_skull_break, {duration = self.actual_break_duration})
+				end
+			end            
 
-                -- If the target is not skull crashed yet, try to immediately CRUSH IT!
-                if not target:HasModifier(self.modifier_skull_crash) then
-                    if RollPseudoRandom(self.insta_skull_break_chance_pct, self) then
-                        target:AddNewModifier(self.caster, self.ability, self.modifier_skull_break, {duration = self.actual_break_duration})  
-                    else
-                        target:AddNewModifier(self.caster, self.ability, self.modifier_skull_crash, {duration = self.skull_break_duration})
-                    end
-                else
-                    -- Otherwise, it was ALREADY CRUSHED! BREAK IT!!!!!!!!!!!! BREAK IT!!!!!!!!!!!!!!!
-                    -- Consume skull crash modifier
-                    target:RemoveModifierByName(self.modifier_skull_crash)
-
-                    -- Apply BREAK!
-                    target:AddNewModifier(self.caster, self.ability, self.modifier_skull_break, {duration = self.actual_break_duration})
-                end
-            end            
-                
-            self.bash_proc = false
-        end
-    end
+			self.bash_proc = false
+		end
+	end
 end
 
 function modifier_imba_abyssal_blade_unique:GetModifierProcAttack_BonusDamage_Magical()
-    if self.bash_proc then
-        return self.bash_damage
-    end
+	if self.bash_proc then
+		return self.bash_damage
+	end
 
-    return nil
+	return nil
 end
 
 
@@ -296,24 +292,22 @@ function modifier_imba_abyssal_blade_bash:CheckState()
 end
 
 function modifier_imba_abyssal_blade_bash:GetEffectName()
-    return "particles/generic_gameplay/generic_stunned.vpcf"
+	return "particles/generic_gameplay/generic_stunned.vpcf"
 end
 
 function modifier_imba_abyssal_blade_bash:GetEffectAttachType()
-    return PATTACH_OVERHEAD_FOLLOW
+	return PATTACH_OVERHEAD_FOLLOW
 end
 
 function modifier_imba_abyssal_blade_bash:DeclareFunctions()
-    local decFuncs = {MODIFIER_PROPERTY_OVERRIDE_ANIMATION}
+	local decFuncs = {MODIFIER_PROPERTY_OVERRIDE_ANIMATION}
 
-    return decFuncs
+	return decFuncs
 end
 
 function modifier_imba_abyssal_blade_bash:GetOverrideAnimation()
-    return ACT_DOTA_DISABLED 
+	return ACT_DOTA_DISABLED 
 end
-
-
 
 -- Modifier responsible for showing that the skull has been broken - Next attack will break it for real!
 modifier_imba_abyssal_blade_skull_crash = modifier_imba_abyssal_blade_skull_crash or class({})
@@ -331,22 +325,22 @@ function modifier_imba_abyssal_blade_skull_break:IsPurgable() return true end
 function modifier_imba_abyssal_blade_skull_break:IsDebuff() return true end
 
 function modifier_imba_abyssal_blade_skull_break:CheckState()
-    local state = {[MODIFIER_STATE_PASSIVES_DISABLED] = true}
+	local state = {[MODIFIER_STATE_PASSIVES_DISABLED] = true}
 
-    return state
+	return state
 end
 
 function modifier_imba_abyssal_blade_skull_break:OnCreated()
-    if IsServer() then
-        local caster = self:GetCaster()
-        local ability = self:GetAbility()
-        local parent = self:GetParent()
-        local particle_break = "particles/item/skull_basher/skull_basher.vpcf"
+	if IsServer() then
+		local caster = self:GetCaster()
+		local ability = self:GetAbility()
+		local parent = self:GetParent()
+		local particle_break = "particles/item/skull_basher/skull_basher.vpcf"
 
-        local particle_break_fx = ParticleManager:CreateParticle(particle_break, PATTACH_ABSORIGIN_FOLLOW, parent)
-        ParticleManager:SetParticleControl(particle_break_fx, 0, parent:GetAbsOrigin())
-        ParticleManager:ReleaseParticleIndex(particle_break_fx)
-    end
+		local particle_break_fx = ParticleManager:CreateParticle(particle_break, PATTACH_ABSORIGIN_FOLLOW, parent)
+		ParticleManager:SetParticleControl(particle_break_fx, 0, parent:GetAbsOrigin())
+		ParticleManager:ReleaseParticleIndex(particle_break_fx)
+	end
 end
